Consolidate configuration-related methods
The new Configuration type (which will eventually be renamed to Config, once
we get rid of the complicated Config and UserConfig stuff) seems to be a better place
for the Dir, userHome, defaultWorkspace, and SetDefaultDirName functions.
