+++ config/user_config.go
@@ -1,13 +1,6 @@
 package config
 
-import (
-	"fmt"
-	"os"
-	"path/filepath"
-	"runtime"
-
-	"github.com/spf13/viper"
-)
+import "github.com/spf13/viper"
 
 var (
 	defaultBaseURL = "https://v2.exercism.io/api/v1"
@@ -63,36 +56,4 @@ func (cfg *UserConfig) Write() error {
 func (cfg *UserConfig) Load(v *viper.Viper) error {
 	cfg.readIn(v)
 	return v.Unmarshal(&cfg)
-}
-
-func userHome() string {
-	var dir string
-	if runtime.GOOS == "windows" {
-		dir = os.Getenv("USERPROFILE")
-		if dir != "" {
-			return dir
-		}
-		dir = os.Getenv("HOMEDRIVE") + os.Getenv("HOMEPATH")
-		if dir != "" {
-			return dir
-		}
-	} else {
-		dir = os.Getenv("HOME")
-		if dir != "" {
-			return dir
-		}
-	}
-	// If all else fails, use the current directory.
-	dir, _ = os.Getwd()
-	return dir
-}
-
-func defaultWorkspace(home string) string {
-	dir := filepath.Join(home, DefaultDirName)
-	_, err := os.Stat(dir)
-	// Sorry about the double negative.
-	if !os.IsNotExist(err) {
-		dir = fmt.Sprintf("%s-1", dir)
-	}
-	return dir
 }
